Committing patch provided by Conal Tuohy to fix the following issue:
If you attempt to deposit an application/zip package into Fedora's SWORD service, the ingestion will fail if one of the files in the package contains spaces in the file name. The pr
oblem is that the filename (from the ZIP) is copied to the foxml:datastream/@id value, which expects an unqualified XML name. The file name needs to be encoded in some way.

more details at http://sourceforge.net/tracker/?func=detail&aid=3339741&group_id=208296&atid=1005125